clementine_utils

Module address

Functions

calculate_taproot_leaf_depths

Calculates the depth of each leaf in a balanced Taproot tree structure. The returned Vec contains the depth for each script at the corresponding index.

create_taproot_address

Creates a taproot address with given scripts and internal key.

is_p2a_anchor

Helper function to check if a TxOut is a P2A anchor.

op_return_txout

Creates an OP_RETURN output with the given data slice.

taproot_builder_with_scripts

A helper to construct a TaprootBuilder from a slice of script buffers, forming the script tree. Finds the needed depth the script tree needs to be to fit all the scripts and inserts the scripts.